requirements.txt
streamlit
python-dotenv
psycopg2-binary
sqlalchemy
google-generativeai
webbrowser
os